Update cost type

+++ schemas/cart.go
@@ -1,8 +1,8 @@
 package schemas
 
 type Cart struct {
-	Id        uint64  `json:"id"`
-	UserId    uint64  `json:"user_id"`
-	ProductId uint64  `json:"product_id"`
-	Cost      float64 `json:"cost"`
+	Id        uint64 `json:"id"`
+	UserId    uint64 `json:"user_id"`
+	ProductId uint64 `json:"product_id"`
+	Cost      uint64 `json:"cost"`
 }
